Handle RT essential BCs and different coefficient modes in saddle point solver
The coefficient for the L2 block can be evaluated in "identity mode", which is
uses the coefficient passed in, or it can be evaluated in "reciprocal mode", in
which the reciprocal is used.

The correct mode depends on the linear system being solved (radiation diffusion
typically requires reciprocal mode, whereas grad-div problems require identity
mode).